// src/lib/generated/Condition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Field } from "./Field";
import type { Op } from "./Op";

/**
 * One test.
 */
export type Condition = { field: Field, op: Op, 
/**
 * The comparison value. Ignored for `IsTrue`/`IsFalse`.
 */
value: string, };